all: majorly refactor to tick at 4MHz instead of 1MHz (fixes audio bugs)
Unfortunately, all devices (on the DMG at least) can be emulated at 1MHz
*except* APU Channel 3 which ticks at 2MHz. For simplicity sake, use the
already-existing 4MHz rate to tick instead of trying to time everything
on 2MHz.

This does introduce a good amount of overhead due a bunch of methods
being called only to immediately exit (for example, the CPU/PPU only
execute 1 time every 4 ticks). This can and should be fixed in the
future, but will be fine for now (other than the fact that CPU usage is
a *little* bit higher than before.)